requirements.txt
# Discord
discord.py==2.6.4

# Database
aiosqlite>=0.19.0

# Configuration
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# External APIs
google-genai>=1.0.0

# Note: aiohttp is already installed via discord.py
